requirements.txt
alembic==1.14.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
Flask==2.2.5
Flask-Limiter==3.10.1
Flask-Migrate==4.1.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
oauthlib==2.1.0
ordered-set==4.1.0
packaging==24.2
psycopg2==2.9.10
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.1.0
rich==13.9.4
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==2.2.3
wrapt==1.17.2
